+++ CHANGELOG.md
@@ -113,6 +113,7 @@ All notable changes to this project will be documented in this file.
 -   #2149 : Fix multi-line expressions in `if` conditions.
 -   #2181 : Allow saving an array result of a function to a slice but raise a warning about suboptimal performance.
 -   #2190 : Fix missing error for list pointer assignment.
+-   Fixed returning strings from functions.
 -   Lifted the restriction on ndarrays limiting them to rank<15.
 -   #2175 : Fix the shape of multi-level containers.